0.9.0.7: 	Begin MIPS megamerge from Thiemo Suifer ("More MIPS-related patches" 	sbcl-devel 2005-04-23)
	This patch merges those pieces which only touch MIPS code, and
	as such I can't really test in a non-trivial way.
	... don't use nl4 (pa-flag) in assembly routines
		Message-ID: <20050422211550.GB10767@hattusa.textio>
	... fix my thinko in the fixnum/sb-xc:fixnum backend immediate-sc
		routine
		Message-ID: <20050422212056.GC10767@hattusa.textio>
	... use linux-nm.  (Other backends maybe should do the same)
		Message-ID: <20050422213247.GE10767@hattusa.textio>
	... lotso mips runtime cleanups
		Message-ID: <20050422224553.GJ10767@hattusa.textio>
	... an untested mips spinlock implementation
		Message-ID: <20050422224830.GK10767@hattusa.textio>
	... mips-assem.S fixes
		Message-ID: <20050422232020.GL10767@hattusa.textio>
	... more runtime fixes
		Message-ID: <20050422233014.GM10767@hattusa.textio>
	... load delay fixes for ldso-stubs
		Message-ID: <20050422233501.GO10767@hattusa.textio>

+++ src/runtime/mips-linux-os.h
@@ -8,7 +8,7 @@ static inline os_context_t *arch_os_get_context(void **void_context) {
     return (os_context_t *) *void_context;
 }
 
-unsigned long os_context_fp_control(os_context_t *context);
+unsigned int os_context_fp_control(os_context_t *context);
 void os_restore_fp_control(os_context_t *context);
 unsigned int os_context_bd_cause(os_context_t *context);
